A-MoreOOP Moved methods out of Duke in the newly added UI, Parser class
