+++ data/fileStorage.txt
@@ -1,0 +1,3 @@
+T | 0 | sayHello to Google
+E | 1 | Google Day | 2nd October
+D | 0 | submit book | 3rd October
