+++ src/main/java/FileStorage/FileManager.java
@@ -1,3 +1,5 @@
+package FileStorage;
+
 import Exceptions.DukeException;
 import Tasks.Deadline;
 import Tasks.Event;
@@ -11,18 +13,25 @@ import java.util.ArrayList;
 import java.util.Scanner;
 
 public class FileManager {
-    static final String SAVE_PATH = "data/fileStorage.txt";
-    public static ArrayList<Task> tasksList = new ArrayList<>();
 
-    public FileManager(File userFile){
+    public static String relativePath;
+    public static ArrayList<Task> tasksList = new ArrayList<>();
+    private String fileName;
+    private String fileDir;
+
+    public FileManager(String fileName, String fileDir){
+        this.fileName = fileName;
+        this.fileDir = fileDir;
+        relativePath = fileDir + '/' + fileName;
     }
+
     /**
      * Saves all tasks into a text file in a specified format
-     * @param tasksList array list of current taks
+     * @param tasksList array list of current task
      * @throws DukeException If file can't be created
      */
-    public static void saveFile(ArrayList<Task> tasksList) throws DukeException, IOException {
-        File savedFile = new File(SAVE_PATH);
+    public void saveFile(ArrayList<Task> tasksList) throws DukeException, IOException {
+        File savedFile = new File(relativePath);
         try {
             if(!savedFile.exists()){
                 savedFile.getParentFile().mkdirs();
@@ -35,30 +44,30 @@ public class FileManager {
     }
     /**
      * Saves all tasks in a specified format
-     * @param taskList array list of current taks
+     * @param taskList array list of current task
      * @throws IOException If file can't be created
      * @throws DukeException If file can't be created
      */
-    private static void saveToFile(ArrayList<Task> taskList) throws DukeException, IOException {
-        File savedFile = new File(SAVE_PATH);
+    private void saveToFile(ArrayList<Task> taskList) throws DukeException, IOException {
+        File savedFile = new File(relativePath);
         FileWriter writer = new FileWriter(savedFile);
         convertToSaveFormat(taskList, writer);
     }
     /**
      * Converts all tasks to a  specified format
-     * @param taskList array list of current taks
-     * @throws IOException If file can't be created
+     * @param taskList array list of current task
+     * @throws DukeException file can't be created
      */
-    private static void convertToSaveFormat(ArrayList<Task> taskList, FileWriter writer) throws DukeException {
+    private void convertToSaveFormat(ArrayList<Task> taskList, FileWriter writer) throws DukeException {
         try {
             for(int i = 0; i< taskList.size(); i++){
                 Task currTask = taskList.get(i);
                 String textFormat = currTask.getType() + " | " +
                         (currTask.getIsDone()?"1":"0") +" | " + currTask.getRawDescription();
                 if(currTask instanceof Event){
-                    textFormat = textFormat + "|" + ((Event)currTask).getAt();
+                    textFormat = textFormat + " |" + ((Event)currTask).getAt();
                 }else if(currTask instanceof Deadline){
-                    textFormat = textFormat + "|" + ((Deadline)currTask).getBy();
+                    textFormat = textFormat + " |" + ((Deadline)currTask).getBy();
                 }
                 writer.write(textFormat + System.lineSeparator());
             }
@@ -69,14 +78,13 @@ public class FileManager {
     }
     /**
      * Load stored TXT file by calling on private loadFromExistingFile function
-     * @throws IOException If file can't be loaded/found
      * @return taskList Array list of tasks containing all stored tasks
      */
-    public static ArrayList<Task> loadFile() throws IOException, DukeException {
+    public ArrayList<Task> loadFile() throws DukeException {
         try {
-            File savedFile = new File(SAVE_PATH);
+            File savedFile = new File(relativePath);
             if(savedFile.exists()){
-                loadFromExistingFile();
+                tasksList = loadFromExistingFile();
             }else{
                 savedFile.getParentFile().mkdirs();
                 savedFile.createNewFile();
@@ -91,8 +99,8 @@ public class FileManager {
      * @throws IOException If file can't be loaded
      * @return taskList Array list of tasks containing all stored tasks
      */
-    private static ArrayList<Task> loadFromExistingFile() throws IOException {
-        File savedFile = new File(SAVE_PATH);
+    private ArrayList<Task> loadFromExistingFile() throws IOException {
+        File savedFile = new File(relativePath);
         Scanner fileScanner = new Scanner(savedFile);
         while(fileScanner.hasNext()){
             String currLine = fileScanner.nextLine();
@@ -102,7 +110,7 @@ public class FileManager {
             case "D":
                 String deadlineDescription = splitString[2].trim();
                 String by = splitString[3].trim();
-                Boolean hasCompletedDeadline = (splitString[1].trim().equals("1"));
+                boolean hasCompletedDeadline = (splitString[1].trim().equals("1"));
                 Deadline newDeadline = new Deadline(deadlineDescription,by);
                 if(hasCompletedDeadline){
                     newDeadline.setAsDone();
@@ -112,7 +120,7 @@ public class FileManager {
             case "T":
                 String toDoDescription = splitString[2].trim();
                 ToDo newToDo = new ToDo(toDoDescription);
-                Boolean hasCompletedToDo = (splitString[1].trim().equals("1"));
+                boolean hasCompletedToDo = (splitString[1].trim().equals("1"));
                 if(hasCompletedToDo){
                     newToDo.setAsDone();
                 }
@@ -122,7 +130,7 @@ public class FileManager {
             case "E":
                 String eventDescription = splitString[2].trim();
                 String at = splitString[3].trim();
-                Boolean hasCompletedEvent = (splitString[1].trim().equals("1"));
+                boolean hasCompletedEvent = (splitString[1].trim().equals("1"));
                 Event newEvent = new Event(eventDescription,at);
                 if(hasCompletedEvent){
                     newEvent.setAsDone();
