+++ src/main/java/Tasks/Task.java
@@ -7,7 +7,7 @@ public class Task {
 
 
     public Task(String description) {
-        this.description = description;
+        this.description = description.trim();
         this.isDone = false;
     }
